+++ tests/test_routes.py
@@ -652,3 +652,23 @@ def test_page_not_found(route):
     response = get(route)
     assert response.status_code == 404
     assert '<title>Page not found | OMEGA iServers</title>' in response.text
+
+
+@pytest.mark.parametrize('route', ['/databases', '/databases/'])
+def test_databases(route):
+    json = get(route).json()
+    assert len(json) == 2
+    assert json['01234'] == {
+        'alias': 'b',
+        'fields': ['timestamp', 'temperature', 'humidity', 'dewpoint'],
+        'max_date': '2021-06-28 21:16:48',
+        'min_date': '2015-01-01 20:29:27',
+        'num_rows': 80,
+    }
+    assert json['56789'] == {
+        'alias': 'f',
+        'fields': ['timestamp', 'temperature1', 'humidity1', 'dewpoint1', 'temperature2', 'humidity2', 'dewpoint2'],
+        'max_date': '2021-06-28 18:53:48',
+        'min_date': '2015-01-01 23:56:47',
+        'num_rows': 80,
+    }